Fix bug on 64-bit Linux that gave dark display
Caused by mismatch in function declaration (float) and definition (float).
Also, logmax, dyn_range and associated functions and associated calculations
all work in floats now.

+++ axes.c
@@ -403,7 +403,7 @@ draw_status_line(void)
     sprintf(s, "%g - %g Hz   %g octaves   %g dB",
 	    min_freq, max_freq,
     	    log(max_freq / min_freq) / log(2.0),
-	    dyn_range);
+	    (double)dyn_range);
     draw_text(s, min_x, max_y + 2, LEFT, BOTTOM);
 
     sprintf(s, "%g pixels per second   %g pixels per octave",
@@ -411,7 +411,7 @@ draw_status_line(void)
 
     draw_text(s, disp_offset, max_y + 2, CENTER, BOTTOM);
 
-    sprintf(s, "%g dB DYNAMIC RANGE", dyn_range);
+    sprintf(s, "%g dB DYNAMIC RANGE", (double)dyn_range);
     draw_text(s, (max_x + disp_offset / 2), max_y + 2, CENTER, BOTTOM);
 
     sprintf(s, "%s WINDOW AT %g HZ", window_name(window_function), fft_freq);
